requirements.txt
adium-theme-ubuntu==0.3.4
attrs==15.2.0
beautifulsoup4==4.4.1
blinker==1.3
ccsm==0.9.12.2
chardet==2.3.0
compizconfig-python==0.9.12.2
cryptography==1.2.3
debtags==2.0
defer==1.0.6
deluge==1.3.12
dirspec==13.10
duplicity==0.7.6
ecdsa==0.13
enum34==1.1.2
Fabric==1.12.0
html5lib==0.999
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
lockfile==0.12.2
lxml==3.5.0
mailgun2==1.1.0
numpy==1.11.0
oauthlib==1.0.3
oneconf==0.3.9
PAM==0.4.2
paramiko==1.17.2
Pillow==3.1.2
piston-mini-client==0.7.5
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pycups==1.9.73
pygame===1.9.1release
pygobject==3.20.0
PyJWT==1.3.0
pyOpenSSL==0.15.1
pyserial==3.0.1
python-apt==1.1.0b1
python-debian==0.1.27
python-harvest-redux==2.0.1
python-libtorrent==1.0.7
pyxdg==0.25
requests==2.11.1
requests-oauthlib==0.7.0
service-identity==16.0.0
six==1.10.0
software-center-aptd-plugins==0.0.0
Twisted==16.0.0
ubuntuone-client-data==14.4
unity-lens-photos==1.0
virtualenv==15.0.3
zope.interface==4.1.3
